app/requirements.txt
django>=3.1.4
django-appconf>=1.0.4
django-imagekit>=4.0.2
mysqlclient>=2.0.3
Pillow>=8.0.1
pytz>=2020.5
gunicorn>=20.0.4
